Create a p4 python lib to abstract shared python functions
This adds a p4 libs folder inside the python path
(see change in the Dockerfile template).

As a start, this moves shared api and github functions and removes
them from the main python scripts.

Other changes:
* Substituted the custom `vprint` method with a simple `logs.append` to
make it easier to abstract things, and since the current scripts are not
expected to be more verbose.
* Made `get_pull_request` also return comments endpoint, for cases where
we want to post a comment on a PR.
* `get_instance_last_commit_date` renamed to `get_last_commit_date` in
order to have a more generic scope.
